+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+* Wrappers for lazy and strict versions of `Data.Map`.
+
 ## 0.1.1
 
 * Initial release on Hackage.